backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.3.2
python-decouple==3.8
django-cors-headers==4.3.1
boto3>=1.34.0
reportlab>=4.0.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
